old v1/requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# cogs\servercount.py: 1
dblpy==0.4.0

# bot.py: 2,6
# cogs\ban.py: 1,2
# cogs\clear.py: 1,2
# cogs\dev.py: 1,2
# cogs\errors.py: 1,2
# cogs\help.py: 1,2
# cogs\info.py: 1,2
# cogs\invite.py: 1,2
# cogs\kick.py: 1,2
# cogs\member_console_logs.py: 1,2
# cogs\nitroboost.py: 1,2
# cogs\ping.py: 1,2
# cogs\prefix.py: 1,3
# cogs\profile.py: 1,2
# cogs\servercount.py: 2,3
# cogs\softclear.py: 1,2
# cogs\stream.py: 1,2
# cogs\support.py: 1,2
# cogs\todo.py: 1,2
# cogs\unban.py: 1,2
# cogs\vcshare.py: 1,2
# cogs\vote.py: 1,2
# cogs\warn.py: 1,2
discord.py==1.4.1
